Introduce one KieSandboxExtendedServicesClient
